Break down/refactor DimensionPickerSidebar
