+++ frontend/src/metabase-types/api/icon.ts
@@ -239,6 +239,7 @@ export type IconName =
   | "sidebar_open"
   | "slack"
   | "slack_colorized"
+  | "sliders"
   | "smartscalar"
   | "snail"
   | "snippet"
