+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/DimensionPickerSidebar.module.css
@@ -8,122 +8,6 @@
 }
 
 .searchInput {
-  height: 2rem;
-  border-radius: 0.5rem;
-}
-
-.item,
-.categoryItem {
-  display: grid;
-  grid-template-columns: 1rem minmax(0, 1fr);
-  gap: 0.5rem;
-  align-items: center;
-  width: 100%;
-  height: 2rem;
-  padding-inline: 0.5rem;
-  border-radius: 0.375rem;
-  color: var(--mb-color-text-primary);
-  font-size: 0.875rem;
-  line-height: 1rem;
-}
-
-.item:hover,
-.categoryItem:hover {
-  background-color: var(--mb-color-background-hover);
-}
-
-.item[data-selected],
-.categoryItem[data-selected] {
-  background-color: var(--mb-color-background-selected);
-  border-radius: 0.5rem;
-  color: var(--mb-color-text-brand);
-}
-
-.item[data-selected]:hover,
-.categoryItem[data-selected]:hover {
-  background-color: var(--mb-color-background-hover);
-}
-
-.categoryRow {
-  align-items: center;
-  gap: 0.25rem;
-  width: 100%;
-}
-
-.categoryItem {
-  flex: 1 1 auto;
-  min-width: 0;
-  border-radius: 0.5rem;
-}
-
-.settingsButton {
-  display: none;
-  flex: 0 0 2rem;
-  width: 2rem;
-  height: 2rem;
-  border-radius: 0.5rem;
-  background-color: var(--mb-color-background-info);
-  color: var(--mb-color-icon-primary);
-}
-
-.categoryRow:hover .settingsButton {
-  display: inline-flex;
-}
-
-.settingsButton.settingsButton:hover,
-.settingsButton.settingsButton[data-expanded="true"] {
-  background-color: var(--mb-color-background-hover);
-  color: var(--mb-color-text-brand);
-}
-
-.settingsIcon {
-  width: 1rem;
-  height: 1rem;
-}
-
-.itemIcon {
-  color: currentcolor;
-}
-
-.itemLabel {
-  overflow: hidden;
-  color: currentcolor;
-  font-size: 0.875rem;
-  font-weight: 500;
-  line-height: 1rem;
-  text-align: left;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.seeAllButton {
-  height: 2rem;
-  padding-inline: 0.5rem;
-  border-radius: 0.375rem;
-  color: var(--mb-color-text-brand);
-  font-size: 0.875rem;
-  font-weight: 600;
-  line-height: 1rem;
-}
-
-.seeAllButton:hover {
-  background-color: var(--mb-color-background-hover);
-}
-
-.categorySelectList {
-  padding-block: 0.375rem 0.625rem;
-  padding-left: 2rem;
-}
-
-.metricSelectInput {
-  min-height: 2rem;
-  height: 2rem;
-  border-color: var(--mb-color-border);
-  border-radius: 0.5rem;
-  font-size: 0.875rem;
-  line-height: 1rem;
-}
-
-.metricSelectInput:not(:focus) {
-  cursor: pointer;
+  border-radius: var(--mantine-radius-md);
+  font-size: var(--mantine-font-size-md);
 }
