+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/components/CategoryButton.module.css
@@ -1,0 +1,69 @@
+.categoryRow {
+  align-items: center;
+  gap: 0.25rem;
+  width: 100%;
+}
+
+.categoryItem {
+  display: grid;
+  grid-template-columns: 1rem minmax(0, 1fr);
+  gap: 0.5rem;
+  align-items: center;
+  flex: 1 1 auto;
+  min-width: 0;
+  width: 100%;
+  height: 2rem;
+  padding-inline: 0.5rem;
+  border-radius: 0.5rem;
+  color: var(--mb-color-text-primary);
+  font-size: 0.875rem;
+  line-height: 1rem;
+}
+
+.categoryItem:hover {
+  background-color: var(--mb-color-background-hover);
+}
+
+.categoryItem[data-selected] {
+  background-color: var(--mb-color-background-selected);
+  color: var(--mb-color-text-brand);
+}
+
+.categoryItem[data-selected]:hover {
+  background-color: var(--mb-color-background-hover);
+}
+
+.settingsButton {
+  display: none;
+  flex: 0 0 2rem;
+  width: 2rem;
+  height: 2rem;
+  border-radius: 0.5rem;
+  background-color: var(--mb-color-background-info);
+  color: var(--mb-color-icon-primary);
+}
+
+.categoryRow:hover .settingsButton {
+  display: inline-flex;
+}
+
+.settingsButton.settingsButton:hover,
+.settingsButton.settingsButton[data-expanded="true"] {
+  background-color: var(--mb-color-background-hover);
+  color: var(--mb-color-text-brand);
+}
+
+.itemIcon {
+  color: currentcolor;
+}
+
+.itemLabel {
+  overflow: hidden;
+  color: currentcolor;
+  font-size: 0.875rem;
+  font-weight: 500;
+  line-height: 1rem;
+  text-align: left;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
