+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/components/CategoryButton.tsx
@@ -1,0 +1,50 @@
+import { t } from "ttag";
+
+import type { DimensionPickerSidebarCategory } from "metabase/metrics-viewer/utils";
+import { ActionIcon, Flex, Icon, Text, UnstyledButton } from "metabase/ui";
+
+import S from "./CategoryButton.module.css";
+
+export function CategoryButton({
+  item,
+  isSelected,
+  canConfigure,
+  isExpanded,
+  onClick,
+  onConfigure,
+}: {
+  item: DimensionPickerSidebarCategory;
+  isSelected: boolean;
+  canConfigure: boolean;
+  isExpanded: boolean;
+  onClick: () => void;
+  onConfigure: () => void;
+}) {
+  return (
+    <Flex className={S.categoryRow} data-expanded={isExpanded || undefined}>
+      <UnstyledButton
+        className={S.categoryItem}
+        data-selected={isSelected || undefined}
+        aria-label={item.name}
+        aria-pressed={isSelected}
+        onClick={onClick}
+      >
+        <Icon className={S.itemIcon} name={item.icon} size={16} />
+        <Text className={S.itemLabel} component="span">
+          {item.name}
+        </Text>
+      </UnstyledButton>
+      {canConfigure && (
+        <ActionIcon
+          className={S.settingsButton}
+          aria-label={t`Configure ${item.name}`}
+          aria-expanded={isExpanded}
+          variant="subtle"
+          onClick={onConfigure}
+        >
+          <Icon name="sliders" />
+        </ActionIcon>
+      )}
+    </Flex>
+  );
+}
