+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/components/CategoryItem.module.css
@@ -1,0 +1,4 @@
+.categorySelectList {
+  padding-block: 0.375rem 0.625rem;
+  padding-left: 2rem;
+}
