+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/components/DimensionButton.module.css
@@ -1,0 +1,42 @@
+.item {
+  display: grid;
+  grid-template-columns: 1rem minmax(0, 1fr);
+  gap: 0.5rem;
+  align-items: center;
+  width: 100%;
+  height: 2rem;
+  padding-inline: 0.5rem;
+  border-radius: 0.375rem;
+  color: var(--mb-color-text-primary);
+  font-size: 0.875rem;
+  line-height: 1rem;
+}
+
+.item:hover {
+  background-color: var(--mb-color-background-hover);
+}
+
+.item[data-selected] {
+  background-color: var(--mb-color-background-selected);
+  border-radius: 0.5rem;
+  color: var(--mb-color-text-brand);
+}
+
+.item[data-selected]:hover {
+  background-color: var(--mb-color-background-hover);
+}
+
+.itemIcon {
+  color: currentcolor;
+}
+
+.itemLabel {
+  overflow: hidden;
+  color: currentcolor;
+  font-size: 0.875rem;
+  font-weight: 500;
+  line-height: 1rem;
+  text-align: left;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
