+++ frontend/src/metabase/metrics-viewer/components/DimensionPickerSidebar/components/MetricDimensionSelect.module.css
@@ -1,0 +1,12 @@
+.metricSelectInput {
+  min-height: 2rem;
+  height: 2rem;
+  border-color: var(--mb-color-border);
+  border-radius: 0.5rem;
+  font-size: 0.875rem;
+  line-height: 1rem;
+}
+
+.metricSelectInput:not(:focus) {
+  cursor: pointer;
+}
